+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Rusoto changes
 
 ## [Unreleased]
-(Please put an entry here in each PR)
+- Add example to Rusoto Logs documentation
+- Add custom dev dependency capability to services crategen
 
 ## [0.33.1] - 2018-08-07